Rename gamed-filter-description to gamed-filter-popover
